login admin

+++ database/migrations/2017_10_08_075406_create_pb_pgs_table.php
@@ -23,8 +23,8 @@ class CreatePbPgsTable extends Migration
             $table->text('description');
             $table->string('profile_picture');
             $table->text('images')->nullable();
-            $table->numeric('height');
-            $table->numeric('weight');
+            $table->integer('height');
+            $table->integer('weight');
 
 
             $table->integer('user_id');
